+++ tests/Module/ParallelRuntimeModuleTest.php
@@ -11,7 +11,7 @@ use PHPUnit\Framework\Attributes\RequiresPhpExtension;
 use PHPUnit\Framework\TestCase;
 use Ray\Di\Injector;
 
-class AsyncParallelBootstrapModuleTest extends TestCase
+class ParallelRuntimeModuleTest extends TestCase
 {
     protected function tearDown(): void
     {
@@ -21,7 +21,7 @@ class AsyncParallelBootstrapModuleTest extends TestCase
     #[RequiresPhpExtension('parallel')]
     public function testContextIsBound(): void
     {
-        $module = new AsyncParallelBootstrapModule('prod-hal-app', 4);
+        $module = new ParallelRuntimeModule('prod-hal-app', 4);
         $injector = new Injector($module);
 
         $context = $injector->getInstance('', Context::class);
@@ -30,12 +30,12 @@ class AsyncParallelBootstrapModuleTest extends TestCase
     }
 
     #[RequiresPhpExtension('parallel')]
-    public function testInstallsAsyncParallelModule(): void
+    public function testInstallsParallelModule(): void
     {
-        $module = new AsyncParallelBootstrapModule('prod-hal-app', 2);
+        $module = new ParallelRuntimeModule('prod-hal-app', 2);
         $injector = new Injector($module);
 
-        // PoolSize should be bound by the installed AsyncParallelModule
+        // PoolSize should be bound by the installed ParallelModule
         $poolSize = $injector->getInstance('', \BEAR\Async\Qualifier\PoolSize::class);
 
         $this->assertSame(2, $poolSize);
@@ -45,13 +45,13 @@ class AsyncParallelBootstrapModuleTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
 
-        new AsyncParallelBootstrapModule('', 4);
+        new ParallelRuntimeModule('', 4);
     }
 
     public function testNonPositivePoolSizeRejected(): void
     {
         $this->expectException(InvalidArgumentException::class);
 
-        new AsyncParallelBootstrapModule('prod-hal-app', 0);
+        new ParallelRuntimeModule('prod-hal-app', 0);
     }
 }